requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=2.0.0
fastapi>=0.110.0
uvicorn>=0.27.1
rich>=13.7.0
python-dotenv>=1.0.1
rapidfuzz>=3.6.1
aiohttp>=3.9.3
python-multipart>=0.0.9
